pyproject.toml
[project]
name = "agent-harness"
version = "1.8.0"
description = "Domain-locked, Temporal-orchestrated agent system for business processes"
requires-python = ">=3.11"
dependencies = [
    "anthropic[vertex]>=0.42.0",
    "temporalio>=1.9.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "pydantic>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pdfplumber>=0.11.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "pypdfium2>=4.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
